Differentiated setup and regular utils
Made a new folder utils in uiplib, and added the setup required methods
in setupUtils.py and others are kept unchanged in utils.py only.

Fixes #224
