+++ uiplib/UIP.py
@@ -3,7 +3,7 @@ import os
 import shutil
 from uiplib.settings import ParseSettings, HOME_DIR
 from uiplib.scheduler import scheduler
-from uiplib.utils import make_dir
+from uiplib.utils.setupUtils import make_dir
 from daemoniker import Daemonizer, send, SIGTERM
